refactor ColorMode out into its own file
